upgrade to syw v0.3
