--- .github/workflows/showyourwork.yml
@@ -1,23 +1,0 @@
-name: article
-
-on:
-  push:
-  workflow_dispatch:
-
-jobs:
-  showyourwork:
-    runs-on: ubuntu-latest
-    name: Build the article PDF
-    concurrency: showyourwork-${{ github.ref }}
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-        with:
-          fetch-depth: 0
-          submodules: recursive
-
-      - name: Build the article PDF
-        id: build
-        uses: ./showyourwork/showyourwork-action
-        env:
-          ZENODO_TOKEN: ${{ secrets.ZENODO_TOKEN }}
