+++ .gitignore
@@ -1,12 +1,14 @@
-# Snakemake temporaries
-.snakemake
-.showyourwork
+# Snakemake and showyourwork temporaries
+**/.snakemake
+**/.showyourwork*
 
 # Showyourwork output files
+/*.pdf
 report.html
 arxiv.tar.gz
-ms.pdf
-dag.pdf
 
-# Showyourwork yaml info 
-showyourwork.yml
+# Figure output
+src/tex/figures/*
+
+# Miscellaneous
+__pycache__
