--- .gitmodules
@@ -1,4 +1,0 @@
-[submodule "showyourwork"]
-	path = showyourwork
-	url = https://github.com/rodluger/showyourwork
-	branch = main
