+++ README.md
@@ -1,6 +1,4 @@
 <p align="center">
-<a href="https://github.com/rodluger/showyourwork">
-</a>
 <br>
 <br>
 <a href="mailto:sarahgthiele@gmail.com?cc=kbreivik@flatironinstitute.org">
@@ -12,35 +10,33 @@
 <a href="https://github.com/katiebreivik/hush/raw/main-pdf/arxiv.tar.gz">
 <img src="https://img.shields.io/badge/article-tarball-blue.svg?style=flat" alt="Article tarball"/>
 </a>
-<a href="https://github.com/katiebreivik/hush/actions/workflows/showyourwork.yml">
-<img src="https://github.com/katiebreivik/hush/actions/workflows/showyourwork.yml/badge.svg" alt="Article status"/>
-</a>
-<a href="https://github.com/katiebreivik/hush/raw/main-pdf/dag.pdf">
-<img src="https://img.shields.io/badge/article-dag-blue.svg?style=flat" alt="Article graph"/>
+<a href="https://github.com/katiebreivik/hush/actions/workflows/build.yml">
+<img src="https://github.com/katiebreivik/hush/actions/workflows/build.yml/badge.svg" alt="Article status"/>
 </a>
 </p>
 
 ## Welcome to _hush_!
 
-This repository contains all of the code necessary to create the results and figures in [Thiele+2021](https://arxiv.org). In this study, we simulate populations of double white dwarf binaries which are orbiting in the frequency band observable by the future space-based gravitational wave detector [LISA](https://www.elisascience.org). Specificially, we investigate the effects of incorporating a [metallicity-dependent binary fraction](https://iopscience.iop.org/article/10.3847/1538-4357/ab0d88) for the first time. 
+This repository contains all of the code necessary to create the results and figures in [Thiele+2021](https://arxiv.org). In this study, we simulate populations of double white dwarf binaries which are orbiting in the frequency band observable by the future space-based gravitational wave detector [LISA](https://www.elisascience.org). Specificially, we investigate the effects of incorporating a [metallicity-dependent binary fraction](https://iopscience.iop.org/article/10.3847/1538-4357/ab0d88) for the first time.
 
 ### Getting started and usage:
 
-To begin, clone the _hush_ repository to a local repository. Dependencies are detailed in [environment.yml](https://github.com/katiebreivik/hush/blob/1eaf321cc5bc97dbc260139181cf2618bc16f833/environment.yml). 
+To begin, clone the _hush_ repository to a local repository. Dependencies are detailed in [environment.yml](https://github.com/katiebreivik/hush/blob/1eaf321cc5bc97dbc260139181cf2618bc16f833/environment.yml). To build the article locally, you should have the latest version of **showyourwork!** installed (`pip install showyourwork`).
 
-There are two levels of complexity at which you can run the paper pipeline for yourself. 
+Running the command `showyourwork` in the `hush` directory will dowload the `results.hdf` data set from [this](https://zenodo.org/record/5722715#.YaA2Sy0ZPOQ) Zenodo and the `FIRE.h5` data set from [this](https://zenodo.org/record/5722451#.YZ152fHMLyg) Zenodo. Once that happens, the scripts to produce each figure will be run: `CEsep.py`, `LISA_SNR.py`, `Mc_vs_dist.py`, `PSD.py`, `form_eff.py`, `lisa_nums.py`, `model_comp.py`, and `sfh_vs_fb.py`.
 
-The first is to download the data created by the pipeline that is required to make the figures. This can be done by running the command `make fast` in the `hush` directory. This will dowload the `results.hdf` data set from [this](https://zenodo.org/record/5722715#.YaA2Sy0ZPOQ) Zenodo and the `FIRE.h5` data set from [this](https://zenodo.org/record/5722451#.YZ152fHMLyg) Zenodo. Finally, the scripts to produce each figure will be run: `CEsep.py`, `LISA_SNR.py`, `Mc_vs_dist.py`, `PSD.py`, `form_eff.py`, `lisa_nums.py`, `model_comp.py`, and `sfh_vs_fb.py`. 
+<!---
+You can also run the entire project pipeline start to finish, which can be done running the command `showyourwork` from the `hush` directory.
 
-You can also run the entire project pipeline start to finish, which can be done running the command `make pdf` from the `hush` directory.
+**!!WARNING!! This process will generate 60+ GB of data on your local disk, and requires ~20 GB of ram to run the simulations.** If you have questions about computational requirements, feel free to reach out to the authors using the contact badge above.
 
-__!!WARNING!! This process will generate 60+ GB of data on your local disk, and requires ~20 GB of ram to run the simulations.__ If you have questions about computational requirements, feel free to reach out to the authors using the contact badge above.
+Running the `showyourwork` command will...
 
-Running the `make pdf` command will...
-- download all the COSMIC + Ananke simulations which lie in [this](https://zenodo.org/record/5722451#.YZ152fHMLyg) Zenodo, and which form the base data for the simulations. These files will be downloaded in your hush directory within `/src/data/`. These can also be downloaded on your own. Once downloaded, the `make pdf` command will not re-download them each time.
+- download all the COSMIC + Ananke simulations which lie in [this](https://zenodo.org/record/5722451#.YZ152fHMLyg) Zenodo, and which form the base data for the simulations. These files will be downloaded in your hush directory within `/src/data/`. These can also be downloaded on your own. Once downloaded, the `showyourwork` command will not re-download them each time.
 - simulate the double white dwarf populations and produce their corresponding data files which again will be in `/src/data/`.
 - run the [scripts](https://github.com/katiebreivik/hush/tree/main/src/figures) to create the paper figure data, which will be contained in a single HDF5 file, `results.hdf`.
 - generate figures and compile the paper, whose text is contained in [ms.tex](https://github.com/katiebreivik/hush/blob/1eaf321cc5bc97dbc260139181cf2618bc16f833/src/ms.tex)
+--->
 
 ### Accessing the paper from this GitHub repo:
