+++ showyourwork.yml
@@ -1,0 +1,49 @@
+version: 0.3.0.dev2
+
+datasets:
+  10.5281/zenodo.5722451:
+    contents:
+      FIRE.h5: src/data/FIRE.h5
+      {% for i in range(15) %}
+      fiducial_10_10_{{i}}.h5: src/data/fiducial_10_10_{{i}}.h5
+      fiducial_11_10_{{i}}.h5: src/data/fiducial_11_10_{{i}}.h5
+      fiducial_11_11_{{i}}.h5: src/data/fiducial_11_11_{{i}}.h5
+      fiducial_12_10_12_{{i}}.h5: src/data/fiducial_12_10_12_{{i}}.h5
+      q3_10_10_{{i}}.h5: src/data/q3_10_10_{{i}}.h5
+      q3_11_10_{{i}}.h5: src/data/q3_11_10_{{i}}.h5
+      q3_11_11_{{i}}.h5: src/data/q3_11_11_{{i}}.h5
+      q3_12_10_12_{{i}}.h5: src/data/q3_12_10_12_{{i}}.h5
+      alpha25_10_10_{{i}}.h5: src/data/alpha25_10_10_{{i}}.h5
+      alpha25_11_10_{{i}}.h5: src/data/alpha25_11_10_{{i}}.h5
+      alpha25_11_11_{{i}}.h5: src/data/alpha25_11_11_{{i}}.h5
+      alpha25_12_10_12_{{i}}.h5: src/data/alpha25_12_10_12_{{i}}.h5
+      alpha5_10_10_{{i}}.h5: src/data/alpha5_10_10_{{i}}.h5
+      alpha5_11_10_{{i}}.h5: src/data/alpha5_11_10_{{i}}.h5
+      alpha5_11_11_{{i}}.h5: src/data/alpha5_11_11_{{i}}.h5
+      alpha5_12_10_12_{{i}}.h5: src/data/alpha5_12_10_12_{{i}}.h5
+      {% endfor %}
+
+dependencies:
+  src/scripts/sfh_vs_fb.py:
+    - src/data/FIRE.h5
+
+  src/scripts/form_eff.py:
+    - src/data/results.hdf
+
+  src/scripts/CEsep.py:
+    - src/data/results.hdf
+
+  src/scripts/lisa_nums.py:
+    - src/data/results.hdf
+
+  src/scripts/Mc_vs_dist.py:
+    - src/data/results.hdf
+
+  src/scripts/PSD.py:
+    - src/data/results.hdf
+
+  src/scripts/LISA_SNR.py:
+    - src/data/results.hdf
+
+  src/scripts/model_comp.py:
+    - src/data/results.hdf
