--- showyourwork.yml.jinja
@@ -1,118 +1,0 @@
-# User configuration; see the docs for details!
-verbose: true
-
-zenodo:
-    - src/data/FIRE.h5: &deposit
-        id: 5722451
-        sandbox: false
-
-{% for i in range(15) %}
-    - src/data/fiducial_10_10_{{i}}.h5:
-        <<: *deposit
-
-    - src/data/fiducial_11_10_{{i}}.h5:
-        <<: *deposit
-
-    - src/data/fiducial_11_11_{{i}}.h5:
-        <<: *deposit
-
-    - src/data/fiducial_12_10_12_{{i}}.h5:
-        <<: *deposit
-    
-    - src/data/q3_10_10_{{i}}.h5:
-        <<: *deposit
-
-    - src/data/q3_11_10_{{i}}.h5:
-        <<: *deposit
-
-    - src/data/q3_11_11_{{i}}.h5:
-        <<: *deposit
-
-    - src/data/q3_12_10_12_{{i}}.h5:
-        <<: *deposit
-    
-    - src/data/alpha25_10_10_{{i}}.h5:
-        <<: *deposit
-
-    - src/data/alpha25_11_10_{{i}}.h5:
-        <<: *deposit
-
-    - src/data/alpha25_11_11_{{i}}.h5:
-        <<: *deposit
-
-    - src/data/alpha25_12_10_12_{{i}}.h5:
-        <<: *deposit
-
-    - src/data/alpha5_10_10_{{i}}.h5:
-        <<: *deposit
-
-    - src/data/alpha5_11_10_{{i}}.h5:
-        <<: *deposit
-
-    - src/data/alpha5_11_11_{{i}}.h5:
-        <<: *deposit
-
-    - src/data/alpha5_12_10_12_{{i}}.h5:
-        <<: *deposit
-{% endfor %}
-
-    - src/data/results.hdf:
-        id: 5722691
-        script: src/figures/pipeline.py
-        title: Applying the metallicity-dependent binary fraction to double white dwarf formation, Implications for LISA, results
-        description: >-
-            This dataset contains data created by post-processing pipeline to produce the figures from Thiele+2021
-        creators:
-            - Thiele, Sarah
-            - Breivik, Katelyn
-            - Sanderson, Robyn E.
-
-dependencies:
-    src/figures/pipeline.py:
-    {% for i in range(15) %}
-        - src/data/fiducial_10_10_{{i}}.h5
-        - src/data/fiducial_11_10_{{i}}.h5
-        - src/data/fiducial_11_11_{{i}}.h5
-        - src/data/fiducial_12_10_12_{{i}}.h5
-        - src/data/q3_10_10_{{i}}.h5
-        - src/data/q3_11_10_{{i}}.h5
-        - src/data/q3_11_11_{{i}}.h5
-        - src/data/q3_12_10_12_{{i}}.h5
-        - src/data/alpha25_10_10_{{i}}.h5
-        - src/data/alpha25_11_10_{{i}}.h5
-        - src/data/alpha25_11_11_{{i}}.h5
-        - src/data/alpha25_12_10_12_{{i}}.h5
-        - src/data/alpha5_10_10_{{i}}.h5
-        - src/data/alpha5_11_10_{{i}}.h5
-        - src/data/alpha5_11_11_{{i}}.h5
-        - src/data/alpha5_12_10_12_{{i}}.h5
-    {% endfor %}
-
-    src/figures/sfh_vs_fb.py:
-        - src/data/FIRE.h5
-
-    src/figures/form_eff.py:
-        - src/data/results.hdf
-
-    src/figures/CEsep.py:
-        - src/data/results.hdf
-
-    src/figures/lisa_nums.py:
-        - src/data/results.hdf
-
-    src/figures/Mc_vs_dist.py:
-        - src/data/results.hdf
-
-    src/figures/PSD.py:
-        - src/data/results.hdf
-
-    src/figures/LISA_SNR.py:
-        - src/data/results.hdf
-
-    src/figures/model_comp.py:
-        - src/data/results.hdf
-
-arxiv_tarball_exclude:
-  - src/data/*.hdf
-  - src/data/*.h5
-  - src/ms.bib
