--- src/.gitignore
@@ -1,15 +1,0 @@
-# Don't track anything...
-*
-
-# ... except the following...
-!figures
-!static
-!.gitignore
-!*.tex
-!*.bib
-
-# ... but never these:
-.showyourwork*
-xstring.tex
-listofitems.tex
-*latexindent*
