+++ src/data/.gitignore
@@ -1,2 +1,3 @@
 # Nothing in this folder should be tracked by git
 *
+!.gitignore
