--- src/figures/.gitignore
@@ -1,7 +1,0 @@
-# Don't track anything...
-*
-
-# ... except the following:
-!.gitignore
-!*.py
-!matplotlibrc
