+++ src/scripts/.gitignore
@@ -1,0 +1,9 @@
+# Don't track figure output
+*.pdf
+*.eps
+*.png
+*.jpeg
+*.jpg
+*.tiff
+*.gif
+*.svg
