+++ src/scripts/PSD.py
@@ -2,25 +2,31 @@ import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 import astropy.units as u
+import paths
 
 
 def func(x, a, b, c, d, e):
     return a + b * x + c * x ** 2 + d * x ** 3 + e * x ** 4
+
 
 model = "fiducial"
 colors = ["#add0ed", "#2b5d87", "#4288c2", "#17334a"]
 Tobs = 4 * u.yr
 
 power_dat_F50 = pd.read_hdf(
-    "../data/results.hdf", key="total_power_DWDs_{}_{}".format("F50", model)
+    paths.data / "results.hdf", key="total_power_DWDs_{}_{}".format("F50", model)
 )
-popt_F50 = pd.read_hdf("../data/results.hdf",  key="conf_fit_DWDs_{}_{}".format("F50", model))
+popt_F50 = pd.read_hdf(
+    paths.data / "results.hdf", key="conf_fit_DWDs_{}_{}".format("F50", model)
+)
 popt_F50 = popt_F50.values.flatten()
 
 power_dat_FZ = pd.read_hdf(
-    "../data/results.hdf", key="total_power_DWDs_{}_{}".format("FZ", model)
+    paths.data / "results.hdf", key="total_power_DWDs_{}_{}".format("FZ", model)
 )
-popt_FZ = pd.read_hdf("../data/results.hdf", key="conf_fit_DWDs_{}_{}".format("FZ", model))
+popt_FZ = pd.read_hdf(
+    paths.data / "results.hdf", key="conf_fit_DWDs_{}_{}".format("FZ", model)
+)
 popt_FZ = popt_FZ.values.flatten()
 
 conf_fit_FZ = (
@@ -94,5 +100,5 @@ plt.xlim(1e-4, 3e-2)
 plt.ylim(1e-38, 5e-34)
 plt.legend(prop={"size": 12}, ncol=2, frameon=False, loc=(0, 1))
 plt.tight_layout()
-plt.savefig("PSD.pdf", dpi=100)
-ax1.set_ylabel(r'PSD [Hz$^{-1}$]', size=15)
+plt.savefig(paths.figures / "PSD.pdf", dpi=100)
+ax1.set_ylabel(r"PSD [Hz$^{-1}$]", size=15)
