+++ src/scripts/matplotlibrc
