+++ src/tex/.gitignore
@@ -1,0 +1,2 @@
+# Don't track TeX temporaries
+*latexindent*
